+++ src/BeeNet/Clients/GatewayApi/IBeeGatewayClient.cs
@@ -74,7 +74,7 @@ namespace Etherna.BeeNet.Clients.GatewayApi
         /// <param name="uid">Uid</param>
         /// <returns>The resource was deleted successfully.</returns>
         /// <exception cref="BeeNetGatewayApiException">A server side error occurred.</exception>
-        Task DeleteTagAsync(int uid);
+        Task DeleteTagAsync(long uid);
 
         /// <summary>Get the list of pinned root hash references</summary>
         /// <returns>List of pinned root hash references</returns>
@@ -137,7 +137,7 @@ namespace Etherna.BeeNet.Clients.GatewayApi
         /// <param name="uid">Uid</param>
         /// <returns>Tag info</returns>
         /// <exception cref="BeeNetGatewayApiException">A server side error occurred.</exception>
-        Task<TagInfoDto> GetTagInfoAsync(int uid);
+        Task<TagInfoDto> GetTagInfoAsync(long uid);
 
         /// <summary>Get list of tags</summary>
         /// <param name="offset">The number of items to skip before starting to collect the result set.</param>
